refactor: split large blob of compiler hook registers into multiple files
