+++ packages/rspack/src/builtin-plugin/html-plugin/index.ts
@@ -1,0 +1,2 @@
+export * from "./plugin";
+export * from "./options";
