+++ packages/rspack/src/builtin-plugin/html-plugin/plugin.ts
@@ -2,22 +2,25 @@ import fs from "node:fs";
 import path from "node:path";
 import {
 	BuiltinPluginName,
-	type JsAfterEmitData,
-	type JsAfterTemplateExecutionData,
-	type JsAlterAssetTagGroupsData,
-	type JsAlterAssetTagsData,
-	type JsBeforeAssetTagGenerationData,
-	type JsBeforeEmitData,
 	type JsHtmlPluginTag,
 	type RawHtmlRspackPluginOptions
 } from "@rspack/binding";
-import * as liteTapable from "@rspack/lite-tapable";
-import { z } from "zod";
 
-import { Compilation } from "../Compilation";
-import type { Compiler } from "../Compiler";
-import { validate } from "../util/validate";
-import { create } from "./base";
+import type { Compilation } from "../../Compilation";
+import type { Compiler } from "../../Compiler";
+import { create } from "../base";
+import {
+	type HtmlRspackPluginHooks,
+	cleanPluginHooks,
+	getPluginHooks
+} from "./hooks";
+import {
+	type HtmlRspackPluginOptions,
+	cleanPluginOptions,
+	getPluginOptions,
+	setPluginOptions,
+	validateHtmlPluginOptions
+} from "./options";
 
 type HtmlPluginTag = {
 	tagName: string;
@@ -27,162 +30,13 @@ type HtmlPluginTag = {
 	toString?: () => string;
 };
 
-export type TemplateRenderFunction = (
-	params: Record<string, any>
-) => string | Promise<string>;
-
-const templateRenderFunction = z
-	.function()
-	.args(z.record(z.string(), z.any()))
-	.returns(
-		z.string().or(z.promise(z.string()))
-	) satisfies z.ZodType<TemplateRenderFunction>;
-
-export type TemplateParamFunction = (
-	params: Record<string, any>
-) => Record<string, any> | Promise<Record<string, any>>;
-
-const templateParamFunction = z
-	.function()
-	.args(z.record(z.string(), z.any()))
-	.returns(
-		z.record(z.string(), z.any()).or(z.promise(z.record(z.string(), z.any())))
-	) satisfies z.ZodType<TemplateParamFunction>;
-
-export type HtmlRspackPluginOptions = {
-	/** The title to use for the generated HTML document. */
-	title?: string;
-
-	/**
-	 * The file to write the HTML to. You can specify a subdirectory here too (eg: pages/index.html).
-	 * @default 'index.html'
-	 */
-	filename?: string | ((entry: string) => string);
-
-	/** The template file path. */
-	template?: string;
-
-	/**
-	 * The template file content, priority is greater than template.
-	 * When using a function, pass in the template parameters and use the returned string as the template content.
-	 */
-	templateContent?: string | TemplateRenderFunction;
-
-	/**
-	 * Allows to overwrite the parameters used in the template.
-	 * When using a function, pass in the original template parameters and use the returned object as the final template parameters.
-	 */
-	templateParameters?: Record<string, string> | boolean | TemplateParamFunction;
-
-	/**
-	 * The script and link tag inject position in template. Use false to not inject.
-	 * If not specified, it will be automatically determined based on scriptLoading.
-	 */
-	inject?: boolean | "head" | "body";
-
-	/** The publicPath used for script and link tags. */
-	publicPath?: string;
-
-	/** Inject a [base](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/base) tag */
-	base?:
-		| string
-		| { href?: string; target?: "_self" | "_blank" | "_parent" | "_top" };
-
-	/**
-	 * Modern browsers support non blocking javascript loading ('defer') to improve the page startup performance.
-	 * Setting to 'module' adds attribute type='module'.
-	 * This also implies 'defer', since modules are automatically deferred.
-	 * @default 'defer'
-	 * */
-	scriptLoading?: "blocking" | "defer" | "module" | "systemjs-module";
-
-	/** Allows you to add only some chunks. */
-	chunks?: string[];
-
-	/** Allows you to skip some chunks. */
-	excludeChunks?: string[];
-
-	/** Allows to control how chunks should be sorted before they are included to the HTML. */
-	chunksSortMode?: "auto" | "manual";
-
-	/** The sri hash algorithm, disabled by default. */
-	sri?: "sha256" | "sha384" | "sha512";
-
-	/**
-	 * Controls whether to minify the output.
-	 * @default false
-	 */
-	minify?: boolean;
-
-	/** Adds the given favicon path to the output HTML. */
-	favicon?: string;
-
-	/**
-	 * If true then append a unique rspack compilation hash to all included scripts and CSS files.
-	 * This is useful for cache busting
-	 * @default false
-	 * */
-	meta?: Record<string, string | Record<string, string>>;
-
-	/** Inject a base tag */
-	hash?: boolean;
-};
-
-const templateFilenameFunction = z
-	.function()
-	.args(z.string())
-	.returns(z.string());
-
-const htmlRspackPluginOptions = z.strictObject({
-	filename: z.string().or(templateFilenameFunction).optional(),
-	template: z
-		.string()
-		.refine(
-			val => !val.includes("!"),
-			() => ({
-				message:
-					"HtmlRspackPlugin does not support template path with loader yet"
-			})
-		)
-		.optional(),
-	templateContent: z.string().or(templateRenderFunction).optional(),
-	templateParameters: z
-		.record(z.string())
-		.or(z.boolean())
-		.or(templateParamFunction)
-		.optional(),
-	inject: z.enum(["head", "body"]).or(z.boolean()).optional(),
-	publicPath: z.string().optional(),
-	base: z
-		.string()
-		.or(
-			z.strictObject({
-				href: z.string().optional(),
-				target: z.enum(["_self", "_blank", "_parent", "_top"]).optional()
-			})
-		)
-		.optional(),
-	scriptLoading: z
-		.enum(["blocking", "defer", "module", "systemjs-module"])
-		.optional(),
-	chunks: z.string().array().optional(),
-	excludeChunks: z.string().array().optional(),
-	chunksSortMode: z.enum(["auto", "manual"]).optional(),
-	sri: z.enum(["sha256", "sha384", "sha512"]).optional(),
-	minify: z.boolean().optional(),
-	title: z.string().optional(),
-	favicon: z.string().optional(),
-	meta: z.record(z.string().or(z.record(z.string()))).optional(),
-	hash: z.boolean().optional()
-}) satisfies z.ZodType<HtmlRspackPluginOptions>;
-
 const HtmlRspackPluginImpl = create(
 	BuiltinPluginName.HtmlRspackPlugin,
 	function (
 		this: Compiler,
 		c: HtmlRspackPluginOptions = {}
 	): RawHtmlRspackPluginOptions {
-		validate(c, htmlRspackPluginOptions);
+		validateHtmlPluginOptions(c);
 		const meta: Record<string, Record<string, string>> = {};
 		for (const key in c.meta) {
 			const value = c.meta[key];
@@ -214,11 +68,11 @@ const HtmlRspackPluginImpl = create(
 		let compilation: Compilation | null = null;
 		this.hooks.compilation.tap("HtmlRspackPlugin", compilationInstance => {
 			compilation = compilationInstance;
-			compilationOptionsMap.set(compilation, c);
+			setPluginOptions(compilation, c);
 		});
 		this.hooks.done.tap("HtmlRspackPlugin", stats => {
-			compilationHooksMap.delete(stats.compilation);
-			compilationOptionsMap.delete(stats.compilation);
+			cleanPluginHooks(stats.compilation);
+			cleanPluginOptions(stats.compilation);
 		});
 
 		function generateRenderData(data: string): Record<string, unknown> {
@@ -386,37 +240,6 @@ function htmlTagObjectToString(tag: {
 	return res;
 }
 
-type ExtraPluginHookData = {
-	plugin: {
-		options: HtmlRspackPluginOptions;
-	};
-};
-
-export type HtmlRspackPluginHooks = {
-	beforeAssetTagGeneration: liteTapable.AsyncSeriesWaterfallHook<
-		[JsBeforeAssetTagGenerationData & ExtraPluginHookData]
-	>;
-	alterAssetTags: liteTapable.AsyncSeriesWaterfallHook<[JsAlterAssetTagsData]>;
-	alterAssetTagGroups: liteTapable.AsyncSeriesWaterfallHook<
-		[JsAlterAssetTagGroupsData & ExtraPluginHookData]
-	>;
-	afterTemplateExecution: liteTapable.AsyncSeriesWaterfallHook<
-		[JsAfterTemplateExecutionData & ExtraPluginHookData]
-	>;
-	beforeEmit: liteTapable.AsyncSeriesWaterfallHook<
-		[JsBeforeEmitData & ExtraPluginHookData]
-	>;
-	afterEmit: liteTapable.AsyncSeriesWaterfallHook<
-		[JsAfterEmitData & ExtraPluginHookData]
-	>;
-};
-
-const compilationHooksMap: WeakMap<Compilation, HtmlRspackPluginHooks> =
-	new WeakMap();
-
-const compilationOptionsMap: WeakMap<Compilation, HtmlRspackPluginOptions> =
-	new WeakMap();
-
 const HtmlRspackPlugin = HtmlRspackPluginImpl as typeof HtmlRspackPluginImpl & {
 	/**
 	 * @deprecated Use `getCompilationHooks` instead.
@@ -465,42 +288,9 @@ HtmlRspackPlugin.createHtmlTagObject = (
 	};
 };
 
-HtmlRspackPlugin.getCompilationOptions = (compilation: Compilation) => {
-	if (!(compilation instanceof Compilation)) {
-		throw new TypeError(
-			"The 'compilation' argument must be an instance of Compilation"
-		);
-	}
-	return compilationOptionsMap.get(compilation);
-};
-
-HtmlRspackPlugin.getHooks = HtmlRspackPlugin.getCompilationHooks = (
-	compilation: Compilation
-) => {
-	if (!(compilation instanceof Compilation)) {
-		throw new TypeError(
-			"The 'compilation' argument must be an instance of Compilation"
-		);
-	}
-	let hooks = compilationHooksMap.get(compilation);
-	if (hooks === undefined) {
-		hooks = {
-			beforeAssetTagGeneration: new liteTapable.AsyncSeriesWaterfallHook([
-				"data"
-			]),
-			alterAssetTags: new liteTapable.AsyncSeriesWaterfallHook(["data"]),
-			alterAssetTagGroups: new liteTapable.AsyncSeriesWaterfallHook(["data"]),
-			afterTemplateExecution: new liteTapable.AsyncSeriesWaterfallHook([
-				"data"
-			]),
-			beforeEmit: new liteTapable.AsyncSeriesWaterfallHook(["data"]),
-			afterEmit: new liteTapable.AsyncSeriesWaterfallHook(["data"])
-		};
-		compilationHooksMap.set(compilation, hooks);
-	}
-	return hooks;
-};
-
+HtmlRspackPlugin.getCompilationOptions = getPluginOptions;
+HtmlRspackPlugin.getHooks = HtmlRspackPlugin.getCompilationHooks =
+	getPluginHooks;
 HtmlRspackPlugin.version = 5;
 
 export { HtmlRspackPlugin };
